requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
